+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+### v1.4.1
+
+Support for React Native.
+
+- Add useFormTag option
+
 ### v1.4
 
 - Use context instead of cloning elements